:art: OmniHttpServiceImpl重命名为BtcOmniServiceImpl :bug: 修复查询纯btc转账时无法正确解析数据

+++ src/main/java/live/lingting/virtual/currency/omni/TransactionByHash.java
@@ -30,7 +30,7 @@ public class TransactionByHash implements Domain<TransactionByHash> {
 	private Long blockTime;
 
 	@JsonProperty("confirmations")
-	private Long confirmations;
+	private BigInteger confirmations;
 
 	@JsonProperty("divisible")
 	private Boolean divisible;
